refactor: update include statements in tracking_object_merger module
