+++ perception/tracking_object_merger/include/autoware_tracking_object_merger/association/solver/mu_ssp.hpp
@@ -12,21 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__SSP_HPP_
-#define AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__SSP_HPP_
+#ifndef AUTOWARE_TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
+#define AUTOWARE_TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
 
-#include "autoware/tracking_object_merger/association/solver/gnn_solver_interface.hpp"
+#include "autoware_tracking_object_merger/association/solver/gnn_solver_interface.hpp"
 
 #include <unordered_map>
 #include <vector>
 
 namespace autoware::tracking_object_merger::gnn_solver
 {
-class SSP : public GnnSolverInterface
+class MuSSP : public GnnSolverInterface
 {
 public:
-  SSP() = default;
-  ~SSP() = default;
+  MuSSP() = default;
+  ~MuSSP() = default;
 
   void maximizeLinearAssignment(
     const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
@@ -34,4 +34,4 @@ public:
 };
 }  // namespace autoware::tracking_object_merger::gnn_solver
 
-#endif  // AUTOWARE__TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__SSP_HPP_
+#endif  // AUTOWARE_TRACKING_OBJECT_MERGER__ASSOCIATION__SOLVER__MU_SSP_HPP_
