+++ perception/tracking_object_merger/src/association/solver/mu_successive_shortest_path_wrapper.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/tracking_object_merger/association/solver/mu_ssp.hpp"
+#include "autoware_tracking_object_merger/association/solver/mu_ssp.hpp"
 
 #include <mussp/mussp.h>
