+++ perception/tracking_object_merger/src/utils/utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/tracking_object_merger/utils/utils.hpp"
+#include "autoware_tracking_object_merger/utils/utils.hpp"
 
 #include "autoware_perception_msgs/msg/shape.hpp"
 #include "autoware_perception_msgs/msg/tracked_object.hpp"
